Extract SearchBar component.
